Revert "Added feature to support delays when showing the loading activity."
